requirements.txt
Django==2.2
django-admin==2.0.1
django-background-task==0.1.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-six==1.0.4
djangorestframework==3.13.1
pytz==2022.1
PyYAML==6.0
screen==1.0.1
six==1.16.0
sqlparse==0.4.2
xlwt==1.3.0
